+++ R/item_reliability.R
@@ -20,12 +20,13 @@
 #' This function calculates the item discriminations (corrected item-total
 #' correlations for each item of `x` with the remaining items) and the
 #' Cronbach's alpha for each item, if it was deleted from the scale. The
-#' absolute value of the item discrimination indices should be above 0.1. An
-#' index between 0.1 and 0.3 is considered as "fair", while an index above 0.3
-#' (or below -0.3) is "good". Items with low discrimination indices are often
-#' ambiguously worded and should be examined. Items with negative indices should
-#' be examined to determine why a negative value was obtained (e.g. reversed
-#' answer categories regarding positive and negative poles).
+#' absolute value of the item discrimination indices should be above 0.2. An
+#' index between 0.2 and 0.4 is considered as "fair", while an index above 0.4
+#' (or below -0.4) is "good". The range of satisfactory values is from 0.4 to
+#' 0.7. Items with low discrimination indices are often ambiguously worded and
+#' should be examined. Items with negative indices should be examined to
+#' determine why a negative value was obtained (e.g. reversed answer categories
+#' regarding positive and negative poles).
 #'
 #' @examples
 #' data(mtcars)
@@ -58,12 +59,12 @@ item_reliability <- function(x, standardize = FALSE, digits = 3) {
     if (standardize) x <- .std(x)
 
     # calculate cronbach-if-deleted
-    cronbachDeleted <- sapply(seq_len(ncol(x)), function(i) cronbachs_alpha(x[, -i]))
+    cronbachDeleted <- vapply(seq_len(ncol(x)), function(i) cronbachs_alpha(x[, -i]), numeric(1L))
 
     # calculate corrected total-item correlation
-    totalCorr <- sapply(seq_len(ncol(x)), function(i) {
+    totalCorr <- vapply(seq_len(ncol(x)), function(i) {
       stats::cor(x[, i], apply(x[, -i], 1, sum), use = "pairwise.complete.obs")
-    })
+    }, numeric(1L))
 
     ret.df <- data.frame(
       term = df.names,